github_issue_adapter/test_requirements.txt
PyYAML
PyGithub
pytest